requirements.txt
openai
langchain
langchain_openai
langchain_community
chromadb
unstructured
python-dotenv
pinecone-client
pymysql
tiktoken
boto3
amazon-textract-caller
pypdf
Pillow